Handle more cases for splitting datasets
Make splitting the experiment transactional by isolating the relevant
operations in a helper service.

Use generic BulkDataSlicerUtils and SingleCellSlicerUtils for slicing
samples in splits which handle all the possible fixed-size encoding
cases and single-cell data slicing.

Add options to the splitExperiment CLI to skip post-processing and
delete the original experiment automatically.
